profiles: uxc: remove obsolete functionality
That profile is known broken since 11/2020 as stated in commit
1d850f40c99e ("profiles,feeds: improve reproducibility"):

 "while doing this update I've noticed broken `profiles/uxc.yml` where
  that `uri` points to repository currently returning 404."

That's probably due to the fact, that `uxc` is no longer separate
project and has been merged into upstream procd project in commit
bb4a4467b1d5 ("uxc: add container management CLI tool") in 7/2020.

prplOS is currently finishing intergration of TR181 orchestrated
container management stack called LCM, thus maintaining of uxc backport
in prplOS is currently out of the project scope and thus it was decided
to remove uxc instead of fixing and maintaining it as a backport.

Closes: PCI-4
Signed-off-by: Petr Štetiar <ynezz@true.cz>