ci: replace job dependencies: with needs:
Currently is not possible to execute manual CI jobs from `cdrouter`
stage if any of the jobs in the previous `run` stage fails.

So fix it by replacing `dependencies:` keyword with `needs:` as this
allows us to execute jobs out-of-order and ignore stage ordering and run
some jobs without waiting for others to complete.

Signed-off-by: Petr Štetiar <ynezz@true.cz>