kernel: netsupport: kmod-sched: explicitly define included modules
Change SCHED_MODULES_EXTRA to an explicit list of modules
instead of taking everything that is not filtered out.
This removes the need of updating the filter each time an extra
sch_*, act_* or similar is added with an own kmod definition.

CONFIG_NET_SCH_FQ_CODEL and CONFIG_NET_SCH_FIFO are already compiled
into the kernel using the generic linux configuration.

CONFIG_NET_EMATCH_TEXT depends on the kmod-lib-textsearch package.

Signed-off-by: Thomas Langer <tlanger@maxlinear.com>
Signed-off-by: Hauke Mehrtens <hmehrtens@maxlinear.com>