hostapd: add config-id and reload_bss patches
This is a combination of the following patches that have been sent to the hostapd mailing list:
https://patchwork.ozlabs.org/project/hostap/patch/20220801110827.237391-2-raphael.melotte@mind.be/
https://patchwork.ozlabs.org/project/hostap/patch/20220801110827.237391-3-raphael.melotte@mind.be/
https://patchwork.ozlabs.org/project/hostap/patch/20220801110827.237391-4-raphael.melotte@mind.be/
https://patchwork.ozlabs.org/project/hostap/patch/20220801110827.237391-7-raphael.melotte@mind.be/

Note that these changes were based on an existing patch in OpenWrt:
https://git.openwrt.org/?p=openwrt/openwrt.git;a=blob;f=package/network/services/hostapd/patches/700-wifi-reload.patch;h=c5ba631a0fc02f70714cb081b42fcf6cb9694450;hb=60fb4c92b6b0d1582d31e02167b90b424185f3a2

However, they only add a config_id parameter for BSSes, not for the
radio.  They also add a new "reload_bss" command, to be able to reload
only one BSS from the control interface.

This patch needs additional changes in the netifd scripts (also from
the upstream OpenWrt project). They are ported over in the next
commit.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>