hostapd:: add config_id to hostapd_set_bss_options()
This change comes from the following commit in upstream OpenWrt:
https://git.openwrt.org/?p=openwrt/openwrt.git;a=commit;h=60fb4c92b6b0d1582d31e02167b90b424185f3a2

It adds a "config_id" parameter to the hostapd configuration that,
when provided, allows to reconfigure only some BSSes.  Note that it
depends on hostapd changes that are not part of this commit (the
changes adding the config_id parameter).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>