ci: tests: generic: 010-firewall-dhcp: make sure fw3 is not interfering
Sometimes we forget to remove fw3 in the config, which leads to various
strange test failures which are hard to reproduce and thus takes time to
hunt down. So make sure, that we don't have fw3 interfering by assuring,
that it's not active in the current system.

Signed-off-by: Petr Štetiar <ynezz@true.cz>
(cherry picked from commit f3e3897df3bff88c657588429dc0b3bce211ddbf)