profiles: prpl.yml: tr181-dhcpv4client: workaround missing nedev-up flag
DHCPv4Client is waiting for wan interface (eth1) to be up before
starting actual DHCP queries and opening proper port in the firewall.

It looks for the state of the WAN interface by doing a netmodel query
waiting for the flag netdev-up to be set on eth1. This flags is in turn
set according to NetDev operstate of eth1: NetDev.Link.[Alias=="eth1"].State
also reflected in /sys/class/net/eth1/operstate.

Turns out on NEC device this operstate remains unkown and this is why
DHCPv4 is never started on the interface.

Until fixed, attempt to workaround it by setting netdev-up flag on WAN
interface once it receives udhcp event and interface has unknown
operstate.

References: PCF-703
Signed-off-by: Petr Štetiar <ynezz@true.cz>