profiles: feed_prpl: bump to cleanup the prplMesh Makefile
Multiple people reported that building prplMesh requires too much RAM
(more than 16GiB). With the make invocation using the '-j' option
without an argument, it will be able to use an unlimited amount of jobs
to build. As a consequence, it could spawn so many jobs that it fills
the RAM completely.

While at it, cleanup all Build steps overrides as it seems, that they're
probably no longer needed and makes the Makefile more ordinary which
should make it easier to maintain, support and future proof as well.

Fixes: PCF-707
Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [commit facelift]