Revert "hostapd: add config-id and reload_bss patches"
This reverts commit 9988aba4fcc7c3173bdc8fc79f8c071040cb804d.

This patch was added for prplMesh, to allow reconfiguring separate
BSSes on a radio, without deauthenticating stations on other BSSes of
the same radio.

It was providing a feature similar to what the
"700-wifi-reload.patch" ([1]) patch provides in upstream
OpenWrt (except that 700-wifi-reload.patch provides much more than
that).

For PPM-2311 ([2]), we want to be able to add and remove BSSes
dynamically. This is provided in upstream OpenWrt by commit
'60fb4c92b6b0d1582d31e02167b90b424185f3a2', which added
"700-wifi-reload.patch". We don't need all the features that this
commit provides, but extracting only the parts that are strictly
needed turns out to be difficult, as multiple commits touching
different components are involved, over multiple months.

Instead, the plan is to backport the complete series ([3]) that
contains commit '60fb4c92b6b0d1582d31e02167b90b424185f3a2', along with
subsequent fixes for it that were added over time.

Since the series provides the feature provided by
"130-config-id-reload-bss.patch" (and much more) remove the patch
entirely now to avoid conflicts in the follow-up commits that will
backport the series.

[1]: https://git.openwrt.org/?p=openwrt/openwrt.git;a=blob;f=package/network/services/hostapd/patches/700-wifi-reload.patch;h=174127df6e117247b2c28eaff0c7f05a928f60c5;hb=HEAD
[2]: https://jira.prplfoundation.org/browse/PPM-2311
[3]: https://patchwork.ozlabs.org/project/openwrt/list/?series=141867&state=%2A&archive=both

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>