ci: sdk.yml: add a possibility to enable CI_PIPELINE_SOURCE rule
This is needed in cases where we're using parent-child pipelines for
merge requests, where we need to set CI_PIPELINE_SOURCE to
`parent_pipeline` otherwise the child pipeline wont start due to very
vague "downstream pipeline can not be created, No stages / jobs for this
pipeline." error. So lets make it possible to set this workflow rule.

Signed-off-by: Petr Štetiar <ynezz@true.cz>