config: make ccache dirs configurable
When using CCACHE, it can be useful to be able to specify a custom
ccache directory outside of the build/staging directories. For example
one might want to share the cache between multiple users, or want to
re-use the same cache after cleaning the build/staging directories.

For this, two new optional config options are introduced: one for the
target builds, and one for the host builds.
If CONFIG_CCACHE is not set, the behavior remains the same as before:
CCACHE_DIR is not exported.
When CONFIG_CCACHE is set, the exported ccache directory is either the
same as before (when the new config options are not set) or the custom
directory specified by the user (when the corresponding new option is
set).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>