ci: 035-multisettings: refactor into separate amxrt instances
We're currently seeing a test run instabilities as the emitter and
receiver are handled by the same amxrt instance.

To apply a config change, multisetting currently restarts the receiver
plugin. In that case it will also restart the emitter plugin. This one
not being setup to have persitency, it would reload the default settings
and thus cause multisetting to react again and restart it with original
profile.

So lets fix it by splitting receiver and emitter into separate amxrt
instances.

References: PCF-583
Reviewed-by: Vincent Harle <vincent.harle@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>