PROVISORY: profiles: x86_64: add workaround for missing prplmesh package
It seems, that current prplMesh package dependencies are relying on
implicit `wpad-basic` package being present in default target's package
set.

Package `wpad-basic` is being set as default for most upstream targets
having wireless interfaces, but it's not the case for example on x86/64
target, where `wpad-basic` is missing and thus `prplmesh` is not being
built as well.

This needs to be fixed properly in the prplMesh package dependencies,
but this would need more time. So for time being lets fix it temporarily
by adding `wpad-basic` into default package set explicitly.

References: PPM-2442
Signed-off-by: Petr Štetiar <ynezz@true.cz>