build: make git sub-modules to fetch configurable
Currently the git protocol downloads all submodules of the target
repository. This can be unwieldy for repositories with a lot of submodules
where only a subset are required in the context of the OpenWrt build.

This change adds a PKG_SOURCE_SUBMODULES variable to configure this
behavior. It takes a space-separated list of submodule paths, or the word
"skip" to disable submodule downloads entirely. The default is to download
all submodules, i.e. preserving current behavior.

References: PCF-1351
Signed-off-by: Karsten Sperling <ksperling@apple.com>
Signed-off-by: Maximilien Baumann <maximilien.baumann@orange.com>
(cherry picked from commit 350d9a34623374b54d2f953917466a5db9ec1dc1)