ci: tests: turris-omnia: adapt cram test
* In 010-network.t:
 - Adpating expected output of ip link.
* In 020-ubus.t:
 - Adapting expected ubus output.
* In 100-prplmesh.t:
 - Adapting expected output of ubus
 - Not  wait for hostapd.wlanx object anymore. Those are loaded when
the rad is up. When enabling the main AP (ie priv) all hostapd.wlan
object are loaded directly. This cannot be used to validate the vaps
activation. Instead check SSID status.
 - Set prplmesh in gateway mode before activate the vaps.
When prplmesh is in mode repeater (ie Agent only) it's supposed to be
connected to a Controller in order to have his FH correctly configured.
If the target is tested alone in mode Agent it can lead to an unexpected
behavior, like the Agent reconfiguring the vaps and causing AP not able
to be activated.

Signed-off-by: Marouan MCHARFI <marouan.mcharfi_ext@softathome.com>