ci: docker: testbed: fix container build using known setuptools
Currently we're using latest and greatest Python3 packages from pip, but
this might result in following breakages with EOLed Python 3.7:

 File "/labgrid/.eggs/setuptools_scm-8.0.3-py3.7.egg/setuptools_scm/_config.py", line 11, in <module>
    from typing import Protocol
 ImportError: cannot import name 'Protocol' from 'typing' (/usr/lib/python3.7/typing.py)

So lets fix it by using known working combinations of pip and
setuptools, do not use python3-setuptools package from Debian.

Moreover it seems, that we need to stick with setuptools version 47.3.2
as this is a last working combination, installation behaves differently
with setuptools version 48.0.0 and labgrid module can't be imported
anymore and needs to be investigated further:

 $ diff -u setuptools-47.3.2.log setuptools-48.0.0.log | grep labgrid
 ...
 -Installed /usr/local/lib/python3.7/dist-packages/labgrid-0.4.1.dev239+g1ca30de-py3.7.egg
 +Installed /usr/lib/python3.7/site-packages/labgrid-0.4.1.dev239+g1ca30de-py3.7.egg
 ...
 +ModuleNotFoundError: No module named 'labgrid'

While at it, invoke pip via '-m pip' to silence "pip is being invoked by
an old script wrapper. This will fail in a future version of pip."
warning.

Fixes: PCF-960
Signed-off-by: Petr Štetiar <ynezz@true.cz>