package: odhcp6c: Add patch to prevent sending of router sollicitations
According to TR181, setting the RSEnable parameter to false should result in no router solicitation messages to be sent from the interface. This seems not the case, the messages are always seen regardless of the parameter value.
To solve this patch odhcp6c to not automatically sends Router Sollicitations as well.

This patch should be properly reworked to make it upstreamable. This progress is tracked in PCF-950

(cherry picked from commit 0f87c3f5a34f85c24e33c174a3baf26df1773056)