version.mk: introduce prplOS Git based versioning
Henceforth, we will utilize the prplOS version throughout the
development cycle, determined by the output of the `git describe`
command. The version will be composed of:

 - The fixed version number of the upcoming release currently in
   development.

 - The abbreviated (8 characters) object hash of the most recent commit.

For instance, with the ongoing development of prplOS version 3.0.0 and
having the Git hash `5a2a69d`, the resulting version would be
`3.0.0-5a2a69d`.

References: PCF-691
Signed-off-by: Petr Štetiar <ynezz@true.cz>