prplOS: target: remove ppp and ppp-mod-pppoe from default packages
We’ve got bug reports, that the new mod-ppp-daemon-v0.5.4 fails to build:

 x86_64-openwrt-linux-musl-gcc -Os <snip>
 In file included from ppp.c:70:
   prplos/staging_dir/target-x86_64_musl/usr/include/pppd/pppd.h:60:10: fatal error: net/ppp_defs.h: No such file or directory
     60 | #include <net/ppp_defs.h>
        |          ^~~~~~~~~~~~~~~~

The problem is that we now have two ppp packages with
different/conflicting staging dir content, `ppp` and `ppp-prpl`, both are
being compiled now, but just `ppp-prpl` ends in the final image.

This can result likely in the following build race:

 Working case:

    ppp               (installs incompatible pppd.h header)
    |
    `ppp-prpl         (installs compatible pppd.h header)
     |
     `mod-ppp-daemon  (builds fine)

 Broken case:

    ppp-prpl          (installs compatible header)
    |
    `ppp              (installs incompatible header)
     |
     `mod-ppp-daemon  (build fails)

So workaround it for now by removing `ppp` and `ppp-mod-pppoe` from
default packages, as we clearly can't have two packages for `ppp` in the
system.

Fixes: PCF-979
Signed-off-by: Petr Štetiar <ynezz@true.cz>