profiles: prpl: add OB-USP-A package and backport libamxrt USP fix
Added OB-USP-A package to feed-prpl with default configuration tailored
for CDRouter's USP certification testing default settings.

To make USP available in the system we need to backport as well fix for
libamxrt which makes libamxrt to only connect to default backend sockets
when the necessary backend is loaded.  So if the USP backend is not
loaded, the plugin won't connect to the USP sockets in /var/run/usp/.

Fixes: PCF-1067
References: PCF-1005
Signed-off-by: Sandesh Mahale <sandesh.mahale@commscope.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [libamxrt fix]
(cherry picked from commit 71646c4dcc68c826b4c1b64896e745255feada6d)