ci: docker: lcm: workaround resolv.conf symlink loop
Container rootfs contains the file /etc/resolv.conf which is a symbolic
link to /tmp/resolv.conf thus LCM agent is not able to set up the
network config correctly as it cannot share the DNS settings with the
container through sharing the file /etc/resolv.conf.

 lxc-start[1]: <ctr-uuid>: utils - ../src/lxc/utils.c:open_without_symlink:1084 - Symbolic link loop - resolv.conf in /usr/lib/lxc/rootfs/etc/resolv.conf was a symbolic link!
 lxc-start[1]: <ctr-uuid>: conf - ../src/lxc/conf.c:mount_entry:2439 - Symbolic link loop - Failed to mount "/lcm/cthulhu/sb_host_data/<ctr>/rootfs/etc/resolv.conf" on "/usr/lib/lxc/rootfs/etc/resolv.conf"

The container can be started without any problem if the container is not
requesting network access or if the file  /etc/resolv.conf is either
removed from the container rootfs or at least set a regular file (not a
symlink) which can then be modified at container startup.

So lets workaround that issue by making /etc/resolv.conf regular empty
file.

References: LCM-737, LCM-746
Suggested-by: Nabil Bizid <nabil.bizid@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>