ci: docker: lcm: workaround updates with different locations
LCM isn't handling correctly Update() when its being done to a different
location and keeps on pointing the previous image blobs when updating to
a different path, when using two different container images versions but
stored on different locations:

 prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64-v1:ynezz-lcm-containers
 prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64-v2:ynezz-lcm-containers

rather than:

 prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64:v1
 prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64:v2

Where the version is v1 and v2 and not ynezz-lcm-containers for both
containers stored in two different subpaths.

 lxc[3860]: BundleVersion: ynezz-lcm-containers
 lxc[3860]: Bundle: prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64-v2
 lxc[3860]: DiskLocation: prpl-foundation/prplos/prplos/prplos/lcm-test-x86-64-v1

So lets fix it by providing container in same path, but tagged with
different versions.

References: LCM-747
Suggested-by: Nabil Bizid <nabil.bizid@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>