ci: cdrouter: fix missing debug-information output
The move command inadvertently included a backslash, resulting in the
shell treating the asterisk as a literal character rather than a
wildcard. This error in the globbing pattern led to the exclusion of
crucial debug-information in the CDRouter job artifacts. This commit
resolves the issue by removing the backslash, ensuring correct pattern
matching and the inclusion of relevant file.

Fixes: 12ca43dc62c4 ("ci: generate coredumps during development")
Signed-off-by: Petr Štetiar <ynezz@true.cz>
(cherry picked from commit a1960c90fd47a5316da55c0ab00337f1eea69c64)