ci: cdrouter: fix missing debug-information output
The move command inadvertently included a backslash, resulting in the
shell treating the asterisk as a literal character rather than a
wildcard. This error in the globbing pattern led to the exclusion of
crucial debug-information in the CDRouter job artifacts. This commit
resolves the issue by removing the backslash, ensuring correct pattern
matching and the inclusion of relevant file.

Fixes: 6cf5635bf50e ("ci: generate coredumps during development")
Signed-off-by: Petr Štetiar <ynezz@true.cz>
(cherry picked from commit 4e151f06160cc8a884713598fb44b333cb4131fb)