build: introduce KERNEL_GIT_BRANCH feature
Using large external Git kernel repositories can be challenging. To
facilitate this, a new KERNEL_GIT_BRANCH feature is added. This feature
allows specifying a Git branch during the cloning process, aiding in
efficiently reaching the desired Git reference in subsequent checkout
step.

This enhancement addresses situation with the DownloadMethod/rawgit
functionality, which performs 'git clone' and 'git checkout'. In cases
of shallow clones, the desired Git reference may not be accessible,
making it necessary to specify the branch separately.

Signed-off-by: Marvin Lu <marvin.lu@wnc.com.tw>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [refactor]