ci: tests: cram: robustify 020-bridging-guest test
We've currently randomly flapping test, likely caused by some race
conditions:

 tr181-bridging: mod-bri - [x]port not added/removed, no bridge name given - (port_ioctl@bridging_ioctl.c:91)
 tr181-bridging: soc_uti - [x]function remove-port failed - (mod_bridging_execute_port_function@bridging_soc_utils.c:257)

so lets fix/workaround it by adding a small sleep between addition
of the new bridge port and enabling of the port.

While at it make the test idempotent, so it can be run in the loop if needed.

Signed-off-by: Petr Štetiar <ynezz@true.cz>