ci: tests: cram: urx851: disable autochannel selection
When this is enable radio are configured to do foreground cac op
to perform wifi dfs. Therefore the radio is not ready to beacon
resulting in a random delay in the activation of VAPs.
Disabling this feature will streamline the VAP activation process.

Signed-off-by: Marouan MCHARFI <marouan.mcharfi_ext@softathome.com>
(cherry picked from commit 0724d8457b00f11484803591988db0bac3eb32ea)