profiles: prpl: feed-prpl update to fetch latest revision of OB-USP-A
Updated prpl profile to point to the feed-prpl branch created for updating OB-USP-A version.
OB-USP-A package updated to fetch the latest version of OB-USP-A from BBF v8.0.0-pre_release branch.
Prominent updates in latest version are listed below:
Added APIs to enable a USP Service to act as a USP Controller (PCF-1073).
Updated INTERNAL_SERVICES_GUIDE.md with example code for a USP Service acting as a Controller.

https://prplfoundationcloud.atlassian.net/browse/PCF-1073

Signed-off-by: Sandesh Mahale <sandesh.mahale@vantiva.com>