ci: urx851-b0-dk: fix properly flapping 010-network.t test
We’re using eth0_1 free interface (there is nothing attached to that
port) for QA of tr181-bridging in the CI in the 020-bridging-guest.t
test. In the test we remove the eth0_1 interface from br-lan bridge, add
it to the br-guest bridge and then attach it into the br-lan bridge back
again. We confirm the validity of changes using bridge link system
command. This test works fine and always passes.

The problem is happening later in the CI job test pipeline, in the
010-network.t test, where we check expected network interface status
using ip link command, there we can spot the regression, where eth0_1
interface is not member of the br-lan bridge anymore.

This happens randomly and for some quite time already, but we should be
only testing for expected test result, not as currently in this case for
the failing test case, so lets fix it by testing against expected test
result and avoid any confusion, so the test fails again only in case of
regression.

References: PCF-1088
Fixes: 16a706939878 ("profiles: prpl: include latest fixes from M2.1/2023 delivery")
Suggested-by: Vincent Harle <vincent.harle@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>