build: add support for checksec.sh based mitigation reports
Checksec is a bash script to check the properties of executables (like
PIE, RELRO, Canaries, ASLR, Fortify Source).

Add an optional build feature, which when enabled is going to create a
JSON report about the state of mitigations in the produced binaries.

Upstream-Status: Submitted [https://github.com/openwrt/openwrt/pull/14569

]
Signed-off-by: Cedric DOURLENT <cedric.dourlent@softathome.com>