ci: cram: urx851-b0-dk-pon: update and sync test cases
Currently the tests for MxL LGP reference device with PON WAN uplink are
a bit outdated, so lets fix that by using symlinks to urx851-b0-dk
tests as they're common and can be shared.

So this results in the new 011-tls.t test and update for 020-ubus.t,
030-wifi.t and 100-prplmesh.t tests.

While at it, fix the expected command outputs for 040-xpon.t test which
have changed in some previous ambiorix updates.

Signed-off-by: Petr Štetiar <ynezz@true.cz>