fixup! profiles: prpl: Bump feed-prpl
Include fixes from https://gitlab.com/prpl-foundation/prplos/feed-prpl/-/merge_requests/251

Update OB-USP-A package to fetch from BBF v8.0.0-pre_release branch the
version of OB-USP-A with fix for failure in registration of non-object
types under Device. (PCF-1084).  Add a patch to remove certain parts of
the data model from the Broker so that they could be registered by a USP
Service.  In obuspa_param_reset.txt remove the setting of
Internal.Reboot.Cause as it is not needed after applying this patch.

https://prplfoundationcloud.atlassian.net/browse/PCF-1084

Signed-off-by: Sandesh Mahale <sandesh.mahale@vantiva.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>