profiles: prpl: OB-USP-A update for fixing failure in registration of non-object types under Device.
Updated prpl profile to point to the feed-prpl branch with following OBUSPA updates.
Add a patch to remove certain parts of the data model from the Broker so that they could be registered by a USP Service.
In obuspa_param_reset.txt remove the setting of Internal.Reboot.Cause as it is not needed after applying this patch.

https://prplfoundationcloud.atlassian.net/browse/PCF-1084

Signed-off-by: Sandesh Mahale <sandesh.mahale@vantiva.com>