profiles: prpl: treewide: fix self killing init scripts
Currently its impossible to stop or restart a lot of HL-API services due
to the usage of `killall` based approach which due to the same naming of
the process/daemons and their init script results into the self killing
init scripts.

So lets fix it by patching temporarily all affected services with fixed
and common process_stop_shutdown init script function which handles it
properly.

While at it refresh the patches so they apply cleanly.

Fixes: PCF-833
Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>