profiles: prpl: add amx-processmonitor v0.2.0
The Process Monitor is a core element of prplOS, designed to oversee
TR-181 services developed with the Ambiorix framework. Its primary
function is to periodically assess the status of TR-181 services linked
to the system bus. This assessment takes place at intervals configured
to suit specific requirements.

In the event of test failures, the Process Monitor, based on its
configuration, can initiate actions such as restarting a single process
or a series of processes. Alternatively, it can trigger a device reboot
to ensure recovering full system functionality if single service restart
is not sufficient. This flexibility contributes to maintaining the
reliability and stability of the prplOS ecosystem, ensuring a seamless
user experience and uninterrupted system performance.

Closes: PCF-897, PCF-1097
Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org> [bump to v0.2.0, tests]