profiles: prpl: obuspa: bump to v8.0.4-prpl-integration3
obuspa, as integrated into prplOS, contains a vendor patch that waits
for the preconditions `Device.DeviceInfo` and
`Device.Time.Status==Synchronized` before attempting to connect to the
USP broker. These preconditions are currently mandatory and cannot be
disabled.

This update fetches the latest version of obuspa from the BBF v8.0.4
prpl-integration3 branch, which adds support for a timeout, controlled
by the parameter `Device.LocalAgent.X_VANTIVA-COM_PreConnectTimeout`.
The default timeout is 60 seconds. This timeout functions as a fallback
mechanism, allowing obuspa to try to connect to the USP broker even if
the preconditions are not met.

References: PCF-1104
Signed-off-by: Sandesh Mahale <sandesh.mahale@vantiva.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>