ci: testbed: fix login timeout by increasing it to 75 seconds
With the platform update to ATH 12.4.0 CS based on 6.1 kernel, we're
experiencing different behavior to previous release ATH 12.2.0 CS based on 5.4
kernel, where the serial console prompt detection takes a bit longer and
sometimes we hit the login timeout of 60 seconds. So lets fix it by
adjusting the timeout slightly to 75 seconds as currently the longest
time experienced was 73.386 seconds.

References: PCF-1177
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>