profiles: prpl: enable KERNEL_WERROR by default
It was enabled upstream in commit b2d1eb717b65 ("generic: 5.15: enable
Werror by default for kernel compile") in 2023-05-12, so year ago and it
proved to be a great static code analysis helper, catching obvious and
hard to spot issues by other means.

Its currently not enabled in prplOS because the config option is
disabled by default if there is user configuration of toolchain options,
which we're doing in security.yml profile.

So lets enable this kernel config option by default explicitly for
improved QA of all prplOS kernels, basically keeping the upstream
defaults in place, but manually.

References: PCF-1171
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>