PROVISORY: ci: testbed: workaround broken Aquantia PHY init in kernel
Currently Aquantia PHY initialization is broken in kernel, leading
either to very long firmware loading time or broken networking.

Tim Hsu proposed workaround, where Aquantia PHY is initialized in
U-Boot, for that `aq_phy_reset` command was added to U-Boot and present
in U-Boot 2016.01.03-prpl release[1]. The workaround resets the Aquantia
PHY in U-Boot and loads the firmware as well, so Linux kernel does not
need to perform this steps.

1. https://gitlab.com/prpl-foundation/platform/qualcomm/u-boot-qca-freedom/-/releases/2016.01.03-prpl

References: PCF-1177
Suggested-by: Tim Hsu <Tim.Hsu@wnc.com.tw>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>