ci: cdrouter: fix unable-to-start-connection-to-usp-agent by enabling NTP sync
Current CDRouter tests are not passing due to the following issue:

 FAIL: The USP agent failed to acknowledge the initial USP message, resulting in the error: "unable-to-start-connection-to-usp-agent".

The failure is occurring because the device is unable to obtain NTP
time. The status of Device.Time is listed as Error_FailedToSynchronize.

Under task PCF-1104, OB-USP-A is configured to withhold attempting
connections to the ACS until essential USP services like DeviceInfo. are
registered, the WAN connection is established, and NTP time is
successfully synchronized. In this scenario, the NTP time has not been
synchronized, which is causing the error.

So lets unblock this by enabling the NTP time synchronization by using
the default/pre-configured prplOS 0.europe.pool.ntp.org and
1.europe.pool.ntp.org NTP servers.

References: PCF-1104
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
(cherry picked from commit 568acaccf8b4fabcdaf66190d74048425d7d7043)