profiles: prpl: disable lighttpd cryptolib
In openwrt 22 the default was:
LIGHTTPD_CRYPTOLIB_NETTLE

In openwrt23 the default is:
LIGHTTPD_CRYPTOLIB_MBEDTLS

But the openwrt23 default gives an error:
Package lighttpd is missing dependencies for the following libraries:
libcrypto.so.3

which can be solved by enabling CONFIG_PACKAGE_lighttpd-mod-mbedtls. Since we do not use crypto in the context of the webui, we can disable linking with a crypto lib by setting CONFIG_LIGHTTPD_CRYPTOLIB_NONE=y and setting CONFIG_LIGHTTPD_SSL to disable.