PROVISORY: prpl-configuration: Add networklayout.json file
It looks like the order of WiFi.Radio.X interfaces is not deterministic. This is caused because the networklayout.json file that is generated has no fixed order of the radio interfaces.

As a workaround, a static networklayout.json is added for the wnc-freedom box.

This issue is related to PPW-20 and should probably be solved by integrating https://gitlab.com/prpl-foundation/prplos/feeds/feed-qca/-/commit/84bd94fac3ab6353cd62ef2164a666fa6eddb842

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>