ci: cdrouter: fix hostname resolutions by flushing DNS relay cache
We're unable to run CDRouter test for CPE's TR-069 functionality due to
the following error during the test init phase:

 FAIL: Did not receive TR-069 Inform RPC from CPE after 120 seconds » no-wan-tr069-inform

Analyzing the logs, issue arise whenever CPE manages to get  a first WAN
address from testbed server instead of CDRouter. With additional unbound traces we can see what happens:

As WAN is up, DNS query for controller.cdroutertest.com  is forwarded to
configured  upstream DSN server (8.8.8.8  and 1.1.1.1)

1.1.1.1 will reply with a SOA  as cdroutertest.com is registered:

 cdroutertest.com.	3600 IN	SOA	ns1.linode.com. support.qacafe.com. 2021000016 14400 14400 1209600 86400

As TTLS is set to 3600 seconds, unbound will cache this as a negative answer for 1 hour.

Latter upstream of CPE is switched to CDRouter, obuspa tries to resolve
controller.cdroutertest.com, but it gets the cached answer as its TTL is
not expired, and no WAN DSN query is done.

So lets fix it by flushing the DNS relay cache during the DUT init
phase, once the network is switched to the CDRouter.

Fixes: PCF-1215
Suggested-by: Vincent Harle <vincent.harle@softathome.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>