profiles: prpl: prplmesh: remove steering vap list
With the current config the steering will be allowed only for the
particular vaps and not for others.

remove the current list, if specific vaps are allocated for steering
then it should be added at the run time.

Closes: PPM-2877

Signed-off-by: rramasamy <rramasamy@maxlinear.com>
(cherry picked from commit f3bca22820b231271f62dc65f7beb524e01ffd60)